requirements.txt
absl-py==0.9.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
filelock==3.0.12
future==0.18.2
idna==2.10
joblib==0.16.0
nltk==3.5
numpy==1.19.0
packaging==20.4
pyparsing==2.4.7
pyspellchecker==0.5.4
regex==2020.6.8
requests==2.24.0
sacremoses==0.0.43
sentencepiece==0.1.91
six==1.15.0
tokenizers==0.8.1rc1
torch==1.5.1
tqdm==4.47.0
transformers==3.0.2
urllib3==1.25.9
